Make buflen a pointer
Make buflen a pointer so that the number of bytes written by xcwd can be
'returned' (and set to 0 if writing was unsuccessful or other errors
occured).
This is useful since the int returned only gives information about xcwd
successfully finding a cwd and not if writing to the buffer succeeded.